Update README.md
This project is a Quiz Web Application designed to evaluate users in various skill sets. The app is built using React and includes core features like a Home screen, a Question screen, and a Report screen. The application follows a structured and scalable approach for state management and utilizes mock APIs for fetching quiz data